Read --character and --trade-hub from the command line. Added a flip appraisal score to order flips by weighted revenue and margin. Update DatabaseAuthDataPersistence to allow for ensuring characters are authenticated. Aggregate market orders by price to reduce number of orders to appraise.

+++ EveStationJanitor.Core/DataAccess/Entities/MarketOrder.cs
@@ -40,9 +40,9 @@ public class MarketOrder
 
     public ItemType ItemType { get; set; } = null!;
 
-    public required int VolumeRemaining { get; set; }
+    public required long VolumeRemaining { get; set; }
 
-    public required int VolumeTotal { get; set; }
+    public required long VolumeTotal { get; set; }
 
     public static OrderRange ParseOrderRange(string range)
     {
